+++ common.js
@@ -19,11 +19,13 @@ function parseCLIArgs (argv) {
       'all',
       'overwrite',
       'strict-ssl',
-      'download.strictSSL'
+      'download.strictSSL',
+      'deref-symlinks'
     ],
     default: {
       'strict-ssl': true,
-      'download.strictSSL': true
+      'download.strictSSL': true,
+      'deref-symlinks': true
     }
   })
 
@@ -187,7 +189,14 @@ module.exports = {
         fs.move(templatePath, tempPath, {clobber: true}, cb)
       },
       function (cb) {
-        fs.copy(opts.dir, appPath, {filter: userIgnoreFilter(opts), dereference: true}, cb)
+        // `deref-symlinks` is the default value so we'll use it unless
+        // `derefSymlinks` is defined.
+        var shouldDeref = opts['deref-symlinks']
+        if (opts.derefSymlinks !== undefined) {
+          shouldDeref = opts.derefSymlinks
+        }
+
+        fs.copy(opts.dir, appPath, {filter: userIgnoreFilter(opts), dereference: shouldDeref}, cb)
       },
       function (cb) {
         // Support removing old default_app folder that is now an asar archive